Global baseline tested and achieved good recall for 1 cluster.
- Turned out that the biggest issue was that the addresses of the
embedings fo do distance calculation with diskann's distanec function
need to be 32-byte aligned. However the blobs have only 16 byte
alignment which caused sigsev. Fixed by changing AVX instrucntion to
allow for unaligned access
- Need to run benchmark to test the qps difference between doing pre
loading a graph in memory and doing it the kv store way is big. If it
is then we might have a big problem
- also need to test for mutiple cluster scenarios