Used raw put instead of affinity set, initialized the thread data correctly for search threads. Compute tasks working, recall 86 because haven't taken into account expanded dist
- Raw put (specifying the subgroup index + shared index) instead
affinity set because affinity set doesn't gurantee that cluster0
cluster1 keys are sharded to different shards
- initialized ssd thread data by having each search/compute thread
call an initialization function provided by SSDIndex. This is correct
and not doing it with omp since the thread ids won't be the same
- Right now recall at around 86 for 3 cluster so the pushing and
receiving of compute tasks seems to be working. However, I forgor to
take into account the expanded distance. With that recall should be 95
for sift 100k