Done with end to end latency analysis for compute query and result
- Bottle neck seems to be from put_and_forget which blocks for 1kus,
resulting in the time from pushing a compute result/query to it being
serialized being ~2kus.
- Another bottle neck is the time from a compute query being pushed
into compute queue to that query actually being computed, around 2k
us. This is because we are doing synchronous io. can fix with io uring
- Communication latency for a batch seems to be 2.5k us locally, will
have to test on cloudlab to see if rdma helps
- serializatio/deserialization is fast, shouldn't be too much of a
problem.